backend/requirements.txt
flask==2.2.5
python-dotenv>=1.0.0
sqlalchemy>=2.0.0